+++ how-to/workspace-platform-starter/client/src/modules/actions/dynamic-menu/shapes.ts
@@ -1,9 +1,0 @@
-/**
- * Options for the dynamic menu actions provider.
- */
-export interface DynamicMenuProviderOptions {
-	/**
-	 * The location of the html page for the popup.
-	 */
-	popupHtml: string;
-}